Split out aggregations tests to seperate file
